pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "speechtotext"
version = "1.0.0"
description = "Voxtral-powered speech-to-text for Windows system tray"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.28.0",
    "pystray>=0.19.0",
    "Pillow>=9.0.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "pyautogui>=0.9.54",
]

[project.scripts]
speechtotext = "speechtotext.main:main"
